+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+v1.0.3
+
+- Changed the memory comparison to use std::memcmp rather than a slow loop
+- Changed string constants to "const char * const" for simplicity and eye
+  toward C++11 compatibility
+- Minor CMakeLists.txt change
+
 v1.0.2
 
 - Revised build options for Windows to enable additional warnings